ISPN-7582 Fix and update rolling upgrade tests
* Improve error message when getting HR codec
* Update README for feature-pack
* Update and fix rolling upgrade tests
