+++ server/integration/testsuite/src/test/java/org/infinispan/server/test/rollingupgrades/HotRodRollingUpgradesDistIT.java
@@ -33,30 +33,7 @@ import org.junit.runner.RunWith;
  */
 @RunWith(Arquillian.class)
 @Category({RollingUpgradesDist.class})
-public class HotRodRollingUpgradesDistIT {
-
-    @InfinispanResource
-    RemoteInfinispanServers serverManager;
-
-    static final String DEFAULT_CACHE_NAME = "default";
-
-    @ArquillianResource
-    ContainerController controller;
-
-    RemoteCacheManagerFactory rcmFactory;
-
-    @Before
-    public void setUp() {
-        rcmFactory = new RemoteCacheManagerFactory();
-    }
-
-    @After
-    public void tearDown() {
-        if (rcmFactory != null) {
-            rcmFactory.stopManagers();
-        }
-        rcmFactory = null;
-    }
+public class HotRodRollingUpgradesDistIT extends AbstractHotRodRollingUpgradesIT {
 
     @Test
     public void testHotRodRollingUpgradesDiffVersionsDist() throws Exception {
@@ -181,26 +158,4 @@ public class HotRodRollingUpgradesDistIT {
         }
     }
 
-    protected RemoteCache<Object, Object> createCache(RemoteInfinispanMBeans cacheBeans) {
-        if (System.getProperty("hotrod.protocol.version") != null) {
-            // we might want to test backwards compatibility as well
-            // old Hot Rod protocol version was set for communication with new server
-            return createCache(cacheBeans, System.getProperty("hotrod.protocol.version"));
-        } else {
-            return createCache(cacheBeans, ProtocolVersion.DEFAULT_PROTOCOL_VERSION.toString());
-        }
-    }
-
-    protected RemoteCache<Object, Object> createCache(RemoteInfinispanMBeans cacheBeans, String protocolVersion) {
-        return rcmFactory.createCache(cacheBeans, protocolVersion);
-    }
-
-    protected RemoteInfinispanMBeans createRemotes(String serverName, String managerName, String cacheName) {
-        return RemoteInfinispanMBeans.create(serverManager, serverName, cacheName, managerName);
-    }
-
-    private Object invokeOperation(MBeanServerConnectionProvider provider, String mbean, String operationName, Object[] params,
-                                   String[] signature) throws Exception {
-        return provider.getConnection().invoke(new ObjectName(mbean), operationName, params, signature);
-    }
 }
