+++ server/integration/testsuite/src/test/resources/config/infinispan/rolling-upgrades-clustered-9.0.xml
@@ -1,4 +1,4 @@
-        <subsystem xmlns="urn:infinispan:server:core:7.0" default-cache-container="clustered-new">
+        <subsystem xmlns="urn:infinispan:server:core:9.0" default-cache-container="clustered-new">
             <cache-container name="clustered-new" default-cache="default">
                 <transport lock-timeout="240000"/>
                 <distributed-cache name="default"
@@ -10,7 +10,7 @@
                                      l1-lifespan="0"
                                      remote-timeout="30000">
                     <transaction mode="NONE"/>
-                    <remote-store cache="default" socket-timeout="60000" tcp-no-delay="true" shared="true" hotrod-wrapping="true" purge="false" passivation="false">
+                    <remote-store cache="default" socket-timeout="60000" tcp-no-delay="true" shared="true" hotrod-wrapping="true" protocol-version="2.0" purge="false" passivation="false">
                         <remote-server outbound-socket-binding="remote-store-hotrod-server"/>
                     </remote-store>
                 </distributed-cache>
