+++ server/integration/testsuite/src/test/resources/config/infinispan/rolling-upgrades-clustered-rest-7.0.xml
@@ -1,4 +1,4 @@
-        <subsystem xmlns="urn:infinispan:server:core:8.0" default-cache-container="clustered-new">
+        <subsystem xmlns="urn:infinispan:server:core:7.0" default-cache-container="clustered-new">
             <cache-container name="clustered-new" default-cache="default">
                 <transport lock-timeout="240000"/>
                 <distributed-cache name="default"
