+++ server/integration/testsuite/src/test/resources/config/infinispan/rolling-upgrades-clustered-rest-8.0.xml
@@ -1,4 +1,4 @@
-        <subsystem xmlns="urn:infinispan:server:core:7.0" default-cache-container="clustered-new">
+        <subsystem xmlns="urn:infinispan:server:core:8.0" default-cache-container="clustered-new">
             <cache-container name="clustered-new" default-cache="default">
                 <transport lock-timeout="240000"/>
                 <distributed-cache name="default"
@@ -10,9 +10,10 @@
                                      l1-lifespan="0"
                                      remote-timeout="30000">
                     <transaction mode="NONE"/>
-                    <remote-store cache="default" socket-timeout="60000" tcp-no-delay="true" shared="true" hotrod-wrapping="true" purge="false" passivation="false">
-                        <remote-server outbound-socket-binding="remote-store-hotrod-server"/>
-                    </remote-store>
+                    <rest-store path="/rest/default" shared="true" purge="false" passivation="false">
+                        <connection-pool connection-timeout="60000" socket-timeout="60000" tcp-no-delay="true" max-connections-per-host="4" max-total-connections="20"/>
+                        <remote-server outbound-socket-binding="remote-store-rest-server"/>
+                    </rest-store>
                 </distributed-cache>
             </cache-container>
             <cache-container name="security"/>
